requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-extensions==2.1.6
django-oidc-provider==0.7.0
future==0.17.1
idna==2.8
pycrypto==2.6.1
pycryptodomex==3.7.3
pyjwkest==1.4.0
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
